Remove parts of //cc we aren't using
The only thing we're using //cc for in Mojo is as part of the
implementation of the surfaces service. Thus we only need the renderers
(in //cc/output), the resource counting system (part of //cc/resources)
and common code to support that. We don't need cc's animation, layers,
trees, input filtering or other systems.

R=tonyg@chromium.org

Review URL: https://codereview.chromium.org/1057283003