Add shell::Tracer object and wire up to --trace-startup on Android
This adds an object to keep track of tracing state and wires it into
the android mojo shell main. When --trace-startup is passed in on the
command line (or via the intent extra args thingy on Android) the shell
enables tracing for itself. It stops tracing when it shuts down cleanly
or 5 seconds elapse. On desktop, this writes into a file named
"mojo_shell.trace" in the CWD. On Android, this writes to
/data/data/org.chromium.mojo.shell/cache/tmp/mojo_shell.trace.

Follow-up patches will patch in trace data from the tracing service into
this file.

R=qsr@chromium.org, viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1067023003