Add MojoGLES2SignalSyncPoint C thunk and impl gpu::ContextSupport
This implements gpu::ContextSupport on top of C entry thunks instead of
downcasting a void* so it can remain binary compatible across different
mojo_shell / app builds and configurations. It also adds a
MojoGLES2SignalSyncPoint entry point since it's currently being used.
The other entry points on gpu::ContextSupport are not used or already
exposed as C thunks. This arguably should be in the sync point extension
table, but since it has to be hand written to deal with callbacks it's
in the control table at least for now. We might want to change this in
the future.

Fixes issue #89

R=alhaad@chromium.org

Review URL: https://codereview.chromium.org/1085743002