Keep the screen on when running mojo shell on a watch.
The default behavior is to close the shell and return to the homescreen 2 seconds after
the last touch. This hack allows us to demo mojo shell on a watch. Eventually we should
figure out how to allow the screen to sleep without closing the shell.

BUG=
R=ppi@chromium.org

Review URL: https://codereview.chromium.org/1094773002