Start a Mojo App by URL.
This is a requested feature per (2nd section, 4th bullet point)
https://docs.google.com/document/d/1J49eJPxIPU3q6za3O0nr3Si3UP1CVusuDyQDmTein-Y/edit#

In this CL, I use the mojo:// scheme to match URIs at the Intent level; this
does not change anything to the way mojo apps handle URIs. Here, apps are
embedded by a window manager, thus are supposed to expose a ViewManagerClient.

I also deleted an obsolete Presubmit rules that clashes with this CL and the
new use of mojo:// URIs.

R=qsr@chromium.org, qsr

Review URL: https://codereview.chromium.org/1095473003