Redraw ganesh_app after becoming backgrounded
This makes it so ganesh_app can redraw itself after MojoShell.apk becomes
backgrounded (so long as the device isn't rotated). This teaches ganesh_app
to redraw itself on input, so we can get it to submit a new frame with
resources, and fixes up the surfaces service to remember the size of the last
submitted frame so it can resize the cc::Display appropriately. The size
stuff is a bit of a hack but it's an improvement over crashing.

Based off of the work in https://codereview.chromium.org/1086553004/

R=etiennej@chromium.org

Review URL: https://codereview.chromium.org/1113473002