Refactor apptest support to work with abstract shells.
This patch introduces Shell abstraction in devtools, with
implementations for Linux and Android.

The logic needed to run shell until completion and retrieve the output is
encapsulated in Shell implementations, so that the apptest runner doesn't need
to reason about platforms or configs anymore.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1120683002