Refactor general apptest support out of existing tools.
This patch carves out a general apptest runner. The runner accepts
arguments for the shell and a (String -> Bool) test to be performed on
the output produced by the shell and returns True iff the test
succeeded.

The existing dart and gtest runners are switched to be based on the
extracted runner.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1115213008