Add mac_tool.py from gyp 0bb67471bca068996e15b56738fa4824dfa19de0
This script is used by the mac build for compiling targets on the host.
It comes from the gyp repo, but is a standalone script independent of
the rest of gyp. This adds the file in the path that the build files
expect it to live in to avoid needing to modify those build rules.

R=cstout@chromium.org

Review URL: https://codereview.chromium.org/1128233002