Adding URLResponse Disk Cache to mojo.
This CL introduces a url response disk cache that allows the shell and content
handlers to access cached response data as file without the need to
stream the full content nor create a new file. It also allows
content handlers to access uncompressed cached archive.

The cache service is loaded internally by the shell because it is used
during application loads.

The android handler is using this service to cache the uncompressed archive of
android applications and the compiled dex files

R=blundell@chromium.org, davemoore@chromium.org

Review URL: https://codereview.chromium.org/1088533003