Add dart packages paths to .gitignore
Dart requires having a directory (or symlink) named 'packages' in
various places. The contents of these should not be considered by source
control. We also have a few paths with the name 'packages' in them that
should be considered by source control like:

/third_party/requests/src/requests/packages
/third_party/requests/src/requests/packages/urllib3/packages

so this adds the dart-required /packages/ paths to the root .gitignore.

R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/1135603002