Fix java handler.
The java handler expects a Mojo-Class entry in the manifest of the
application. The latest dx tool strip the manifest to its bare minimum
and removed everything we added. This change the build system to add the
entry to the manifest after running dx.

R=blundell@chromium.org

Review URL: https://codereview.chromium.org/1126393005