Moterm part 6: Add moterm_example_app.
This is an example app that is itself embeddable, and in turn embeds a
single moterm. It uses that moterm to prompt the user about what to
connect to (e.g.: mojo:echo_terminal,
mojo:netcat?host=localhost&port=80), and then connects the moterm to
whatever the user enters. After that connection is done (typically
terminated by ^D on a line by itself), it prompts the user again, etc.

R=erg@chromium.org

Review URL: https://codereview.chromium.org/1137033005