Android handler: Fix name collision.
The android handler was using the android framework to create temporary
files. This creates predictible names and ends up creating 2 libraries
with the same name. This is then an issue with ldopen that will not open
those independently.

This CL changes the implementation to create names that will not be
predictible.

R=blundell@chromium.org

Review URL: https://codereview.chromium.org/1133283002