Add ServeLocalDirectory to the shell abstraction.
This patch adds a ServeLocalDirectory method to the shell abstraction,
allowing to make content of any directory on the host machine available
to the shell under a url. On both platforms it means serving the
content using a local http server, on Android it additionally involves a
reverse port forwarding set up for the device.

Note that the specific implementation of the server logic is still up
in the air at this point - this just introduces an API in front of what
we have today.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1139053005