Fix double vision of Android shell output in Shell::Run().
"Diplopia, commonly known as double vision, is the simultaneous
perception of two images of a single object that may be displaced
horizontally, vertically, diagonally or rotationally in relation to each
other" (Wikipedia)

In this particular case, in Android implementation of Shell::Run() we
are both streaming the logcat produced by the shell to stdout as well
as passing there the actual proper output of the shell run.

In this patch we filter out of the logcat the native log printouts which get
onto shell stdout.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1149483002