Introduce AuthenticatingURLLoader
This CL introduces AuthenticatingURLLoader, which is a wrapper around the
network's URLLoader that supplies OAuth2 authentication via the
AuthenticationService in case of need. It is inspired by
https://codereview.chromium.org/1017713002/.

In particular, this CL does the following:
- Introduces the AuthenticatingURLLoader interface and impl
- Changes the shell's NetworkFetcher to use AuthenticatingURLLoader

On a request, the AuthenticatingURLLoader operates as follows:
- Starts a URLLoader for the request
- If 401 is returned, connects to the AuthenticationService, obtains a username
  to use to obtain a token, and obtains the token.
- Tries the request again with the token as authorization.
- If 401 is returned again, tells the AuthenticationService to clear the token
  for the given username, obtains a fresh token, and tries the request again
  with *that* token.
- If the request fails with the fresh token, gives up.

Clients must supply an AuthenticationService for the AuthenticatingURLLoader to
use via AuthenticatingURLLoaderFactory. This organization ensures that when the
AuthenticationService displays the identity of the app that wants permissions
to the user, it will be the app that requested an AuthenticatingURLLoader
rather than the AuthenticatingURLLoader app itself.

Followup work is needed for the following:
- Using caching within AuthenticatingURLLoader to reduce the frequency with
  which it interacts with the user.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1148013002