Use a deque instead of a list in AwakableList.
In Chrome, AwakableList::Add causes ~32,000 heap allocations when
loading The Verge (and the same number of frees in AwakableList::Remove).
After this change, this drops to 0 (in both), measured in a debug build.

In terms of microbenchmarks, MojoChannelPerfTest.ChannelPingPong doesn't
appear to change, but MojoChannelPerfTest.ChannelProxyPingPong shows some
improvement. However, I would take these results with a grain of salt
since there appears to be a fair bit of variance in these.

From 10 runs, these are the mean (and standard deviation) of a couple of
tests in milliseconds (release build).

                                  std::list     std::deque
IPC_Channel_Perf_50000x_12         1043 (17)     1028 (14)
IPC_ChannelProxy_Perf_50000x_12    2973 (371)    2199 (306)

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1150873003