Use a vector<> instead of a deque<> in AwakableList.
The previous change to switch to a deque<> caused a performance
regression in CorePerftest.MessagePipe_CreateAndClose. This is because
GCC's deque implementation does a heap allocation when creating an empty
deque. vector<> does not, and instead does the allocation on the first
element insertion. Using a vector keeps all the benefits of using a
deque, but shifts the heap allocation from construction to first-use.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1158353002