Move magic rewriting of map-origin arguments to android_mojo_shell.
The logic that magically spawns a local server for each local destination
passed as map-origin doesn't belong in a low level shell abstraction. For
instance, it should be possible to push some files to the device and use
--map-origin with a local device file system path.

This patch refactors this logic and moves it to android_mojo_shell.py. We can
imagine moving it back to devtools in some form, but not as an obligatory part
of StartShell().

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1142223002