Refactor the apptest test list logic into devtools.
This patch extracts the apptest list logic into run_apptest function in
devtools that accepts the shell and the list of tests to be run.

This allows to share this logic between consumers, who just need to
configure the shell to use the runner (see the changes in
apptest_runner.py).

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1166563003