Make mojo::embedder::ConnectToSlave() a bit more "correct".
Now it actually establishes a |Channel| and an initial message pipe.
(The implementation is in mojo::system::IPCSupport.) The old
IPCSupport::ConnectToSlave() was renamed ConnectToSlaveInternal() and
made private.

ConnectToMaster() needs to have the "same" thing done to it, at which
point I'll be able to test the two together.

(Also to do: Make the management of the channels created by
|ConnectToSlave()| more sane.)

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/1166303002.