go/build rules: inherit GOPATH from the environment
This CL defines the MOJO_GOPATH environment variable which
should be used the same way as GOPATH if you want to add
external Go dependencies. This variable makes it easier to
play with the Go bindings in case you need to use a custom Go
library.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1172523002.