Fix tracing_app occasionally failing to retrieve any traces.
Tracing app wasn't verifying in StopTracing() that the message pipes we wait
for were not already closed. When that was the case, we were passing invalid
handles to WaitMany(), and it was returning immediately with
MOJO_RESULT_INVALID_ARGUMENT, blocking the thread in a loop over subsequent
WaitMany() calls and preventing reception of the trace data.

This patch removes all controllers that already closed their message pipes  at
the beginning of StopTracing(). This fixes #225.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1171053002.