Remove use of base::Pickle from message_pipe_perftest.cc.
It doesn't really use much in the way of the capabilities of
base::Pickle, and we'd like to minimize (and ultimately remove)
//mojo/edk's dependencies on //base.

(Note: This theoretically affects the measured perf results, since
very slightly less data will be transmitted. However, the difference is
probably so small as to not be measurable for practical purposes.)

R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1183803002.