Exclude android_stack_parser from pylint presubmit checks.
Pylint picks up plenty of stylistic and other issues in this code which we
adopted from elsewhere. We might decide it's alright to diverge from the
original (making importing changes harder) and fix them, but for now this patch
disables linting on these files, matching what we hard before the move in
https://codereview.chromium.org/1186113002/.

R=etiennej@chromium.org

Review URL: https://codereview.chromium.org/1176153004.