Make the various Dispatcher implementations have a Create() static factory method.
I'd like to get rid of:

scoped_refptr foo(new Foo());

(except possibly something like that in the static factory method).

Rationale: I'd like to move to a refcounted class that begins the
refcount at 1, in which case there needs to be a distinction between
adopting a ref from a Foo* (in the case of new Foo()) and taking a ref
from a Foo* (the usual case when you have a Foo*). (Such a change is
good for two reasons: refcount != 0 will be an invariant for live
objects, and it'll avoid a (probably atomic) increment for each
refcounted objected.)

The easiest/sanest way to do this is to force everyone to go through a
static factory method (which can be changed once we change the
refcounted base as above).

(Note: SharedBufferDispatcher already has a slightly-odd Create(), so it
gets a new static CreateInternal() instead, which Create() also calls.)

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/1218883003.