Convert WeakInterfacePtr (used by WeakInterfacePtrSet) to use set_connection_error_handler().
Also put WeakInterfacePtr in an |internal| namespace, since its name and
functionality are rather odd and it's not used by anything other than
WeakInterfacePtrSet.

And also add a TODO -- it seems to me that WeakInterfacePtrSet
can/should be refactored entirely. This isn't hard, but there currently
no tests for it. :-(

R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1229693003 .