Make WeakBindingSet/WeakBinding use set_connection_error_handler() and not refer to ErrorHandler.
Make WeakBinding internal, since it's a rather odd class (not used
outside of WeakBindingSet). (Like WeakInterfacePtr, it should no longer
be needed, and WeakBindingSet should be refactored.)

Remove WeakBindingSet::set_error_handler() altogether. As an API, it's
rather poor -- it'll inform the error handler that some binding in the
set had a connection error -- and any use of it is suspect.

Indeed, the one use of it by HttpServerImpl was strange and inexplicable
-- on a connection error from the bindings_ (a WeakBindingSet), it went
and looked for connection errors in handlers_. So I removed that use.

(Conceivably, AddBinding() should have the ability to take a connection
error handler Closure, but this is probably over-complicated. With
connection error callbacks, you'd probably be better off managing the
bindings yourself.)

R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1230653002 .