Add a MOJO_NOT_THREAD_SAFE annotation for use in headers.
It's functionally equivalent to MOJO_NO_THREAD_SAFETY_ANALYSIS, but
semantically different: it declares that a function/method is not
thread-safe, so the caller has to be careful to adhere to whatever the
contract is.

E.g., Init() methods are labelled using this in the .h files instead of
using MOJO_NO_THREAD_SAFETY_ANALYSIS in the .cc files (since the caller
must usually adhere to a restriction along the lines of "must call
Init() before calling any other method").

R=kulakowski@chromium.org

Review URL: https://codereview.chromium.org/1231383006 .