Dynamically forward dart observatory ports when running AndroidShell.
This patch adds a watcher thread that parses device logcat looking for
messages indicating the dart observatory being set up by Sky or Dart
content handlers:

  "Observatory listening on http://127.0.0.1:(\d+)"

When such message is found, we find an available port on the host
machine, and forward it to the observatory port on the device, making
the observatory available at the host. At this point we print:

  "Dart observatory available at the host at http://127.0.0.1:%d".

Fixes domokit/devtools#3.

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1239813002 .