Extract the shell runner into devtools.
In this patch, mojo/tools/mojo_shell.py is moved to devtools. Default
output paths corresponding to a regular Mojo checkout are inferred, so
that it continues to work out-of-the-box there.

In the next step we will make it easier (in particular, possible) to use
the runner outside of the regular Mojo checkout too (allow to indicate
some paths by hand, allow to run against already installed device Mojo
Shell, etc.)

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1242453003 .