Allow the shell runner to run outside of Mojo checkout.
This patch allows to experiment with the shell runner outside of a
mojo checkout. Currently this works only for Android, and only if the
shell is already installed on the device.

In the next step we can integrate download_shell_binary to pull down the
shell binary from the cloud if needed.

Fixes domokit/devtools#26.

R=tonyg@chromium.org

Review URL: https://codereview.chromium.org/1248533002 .