Add getPid() support to Dart on Mojo.
In the current non-sandboxed environment we just call the native getpid() function. When we run in the sandbox in the future the getpid() function will probably return a fake pid such as 1. In this case we need to decide what behavior we want Dart's get pid to have.

BUG=#341
R=johnmccutchan@google.com, jln, johnmccutchan

Review URL: https://codereview.chromium.org/1266683002 .