Track connections between processes in MasterConnectionManager.
This allows it to respond with "reuse connection".

Note that if processes P1 and P2 are being connected, possibly multiple
times simultaneously (e.g., if P3 is currently proxying between P1 and
P2 and wants to remove itself by directly connecting P1 and P2, and
ditto for P4), then it is always the first call to Connect() in each
process that gets the platform handle.

R=jam@chromium.org

Review URL: https://codereview.chromium.org/1267773006 .