Devtools: refactor shell configuration logic.
This patch splits the shell configuration logic into part that computes the
final list of options (shell_config) and the part that configures the shell
abstraction accordingly.

This becomes needed as information that affects a shell run (script
command-line arguments and inferred checkout paths), and we would like
to have another (configuration file).

This patch also includesthe logic that infers default paths in Chromium-like
checkouts in the common configuration logic, so that it now applies to
`mojo_test` as well as to `mojo_run`.

R=qsr@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1256323011 .