Save Entry on background thread.
Saving the entry can take 280 ms if it is done concurrently to the
network service saving its cache to disk. Moving this to a background
thread allows to launch the mojo application without waiting for the
disk. It is fails, it will re-download the application on the next run.

R=ppi@chromium.org

Review URL: https://codereview.chromium.org/1272183003 .