Rename ChannelEndpoint::ChannelState to ...::State.
Mainly, the reason is separate "being attached" (i.e., having a non-null
Channel) from the running state.

PAUSED means that the endpoint is alive (on "this" side -- e.g., the
handle to the message pipe hasn't been closed) but not sending messages
to the Channel (such messages are enqueued and will be sent on
transition to RUNNING).

RUNNING means that the endpoint is alive and sending messages to the
Channel (which must thus be non-null; we must also have local and remote
IDs).

DEAD means that the endpoint is permanently dead -- it will never have a
Channel again, and messages do not need to be enqueued.

In the future, there'll be situations in which we are PAUSED but
attached (i.e., have a Channel) in some way.

R=jam@chromium.org

Review URL: https://codereview.chromium.org/1282423004 .