Fix race condition in serializing data pipe consumer handles with remote producers.
In doing so, the Channel replaces the ChannelEndpoint's client. But it
can't ensure that the old client's OnReadMessage() won't get called.

Thus RemoteProducerDataPipeImpl::OnReadMessage() should return false in
that case (and then the ChannelEndpoint will retry).

This is entirely analogous to what we do for
MessagePipe::OnReadMessage(). See also the comment in
ChannelEndpoint::OnReadMessageForClient().

It should be harmless (I think) to return false even if it's known that
the producer is genuinely closed, since
ChannelEndpoint::OnReadMessageForClient() would just notice in the next
iteration that it doesn't have a client.

(The way this bug arises in practice is if a data pipe consumer handle
is passed over a message pipe that already has a proxy/relayer in the
middle, and simultaneously data is being written to the producer. At the
proxy, when the message is deserialized, a consumer w/remote producer is
created then immediately replaced by a relayer. Incoming data messages
may get dropped during this replacement if they are sent to the old
client -- the consumer w/remote producer.)

R=jamesr@chromium.org
BUG=Fixes #373

Review URL: https://codereview.chromium.org/1291213002 .