Mojom compiler: Improves error message by giving misspelled identifier.
Instead of printing a meaningless Python object ID and a cryptic Python error print "unrecognized kind" and the type-spec for the unrecognized kind.

This came up for me when I was writing a longer .mojom file and I had misspelled the name of one of my structs when I was attempting to include the struct in a union. The compiler told me something was wrong but I had no idea which identifier I had misspelled.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1294043002 .