Buffer shell stdout on Android by line when redirected to fifo.
Android shell redirects its stdout to a fifo so that it can be streamed
by `mojo_run` to the host terminal. This patch uses setvbuf to ensure
that the output is buffered by line, matching behavior one expects when
running attached to an interactive terminal (e.g. on Linux).

Fixes #380.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1293173002 .