Return partial output when shell.run_and_get_output time limit is hit.
This patch fixes run_and_get_output() of the Linux shell abstraction so
that if the process times out and is forcefully terminated, any output
produced by it before termination is still returned to the caller,
aligning it with the Android shell abstraction.

We also grow simple tests for run_and_get_output(), using 'echo' and
'cat' instead of the real shell binary.

R=etiennej@chromium.org

Review URL: https://codereview.chromium.org/1301613003 .