Forward deps to android_apk finalizer target
The android official builder is failing with this error:

Running gn 'gen' '--check' 'out/android_Official' <omitted>
ERROR Input to target not generated by a dependency.
The file:
  //out/android_Official/mojo_shell-official.keystore
is listed as an input or source for the target:
  //shell:mojo_shell_apk__create__finalize
but this file was not generated by any dependencies of the target. The target
that generates the file is:
  //shell:download_keystore

If you have generated inputs, there needs to be a dependency path between the
two targets in addition to just listing the files. For indirect dependencies,
    the intermediate ones must be public_deps. data_deps don't count since they're
    only runtime dependencies. If you think a dependency chain exists, it might be
    because the chain is private. Try "gn path" to analyze.

The //shell:mojo_shell_apk target has a dep on :download_keystore, but this dep
isn't forwarded into the finalizer rule. This forwards all deps on the create_apk
target into the finalize target which is overbroad (it includes many things that
the finalize step does not need) but includes the correct dependency as well.

R=ppi@chromium.org

Review URL: https://codereview.chromium.org/1294233003 .