Improve android stack parser
There was several issues with the Android stack parser that prevented it
to work correctly in a number of cases. The two cases fixed here:
 - When a mojo app is retrieved with arguments
(dart_content_handler.mojo?strict=true), we want to ignore ?strict=true
when looking for a match on the local filesystem.
 - There were cases when paths were not normalized correctly (for example,
double-slashes), preventing the libraries to match.

This fixes domokit/devtools#23

R=ppi@chromium.org, ppi

Review URL: https://codereview.chromium.org/1306603002 .