Remove erroneous DCHECK() in RemoteProducerDataPipeImpl::Disconnect().
Also add a test for this case. Unfortunately, the DCHECK() only fails
racily, so we can only test this case racily. (Without this fix, the
test should fail flakily. There's a sleep to encourage the race; without
it, it should still fail flakily but perhaps not measurably so.
Unfortunately, waiting for the producer to be known to be closed puts
the system in a different state, so the DCHECK() wouldn't fail.)

R=jamesr@chromium.org
BUG=Fixes #387

Review URL: https://codereview.chromium.org/1311143004 .