Dart: Fix rebuild problem
Ninja was recording an mtime of 0 for some of the symlinks created by
the dart_pkg_helper rule, which was causing lots of stuff to be
unnecessarily rebuilt.

Instead of depending explicitly on these symlinks, dependent rules now
depend on the stamp file created by dart_pkg.py, which is a regular
file.

This change also creates a Dart package for the interfaces used by
Dart's apptests.

BUG=
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1320723002 .