Allow multiple connections to view_manager
For each connection to view_manager, create a different set of service
instances. This means that each connection to view_manager gets a new and
different Native Viewport to draw onto.

Given that only window managers use this interface, and that the number of
connections was previously capped at one, there should not be any visible
change of behavior given the current ecosystem.

A companion CL modifying the behavior of WindowManager to take advantage
of this change will follow (see issue #382).

This fixes #374.

R=ppi@chromium.org

Review URL: https://codereview.chromium.org/1299643003 .