Benchmark dart app initialization.
This patch allows the dart tracing helper to record events before the
tracing service connects to it, so that dart app initialization can be
traced and benchmarked. The events are enqueued for 1s and dropped on
the floor if the tracing service does not connect.

It also updates the existing dart benchmark to measure the startup time
of dart application.

Fixes #414.
Fixes #415.

This allows us to track progress on #405 .

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1318923005 .