Enabling 64-bit mojo shell to launch 32-bit child to handle nonsfi content.
This patch allows the nonsfi content handler to be used from the mojo
shell in two ways. First, it can be built for a purely 32-bit system
(--target-cpu=x86), and run without "--enable-multiprocess". Alternatively,
the nonsfi content handler can be run from a 64-bit version of the mojo
shell, as long as "--enable-multiprocess" is used. This will fork/exec a
32-bit child whenever the nonsfi content handler is detected.

BUG=https://github.com/domokit/mojo/issues/396
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/1341873002 .