Offline By Default
Update the shell to try to run applications from the cache
before trying to download those from the web.

If an application is thus run from the cache, the shell
will asynchronously connect the server to update the
application for the next run, if needed.

BUG=https://github.com/domokit/mojo/issues/363
R=ppi@chromium.org

Review URL: https://codereview.chromium.org/1276073004 .