Add invalidation to url_response_disk_cache.
Entry in the disk cache are invalidated when requested offline. They can
be revalidated by caching a new version or explicitely revalidating the
entry.

This is necessary because if a downloaded application is buggy and crash
the shell, it will never be updated and the user would be stuck in a
crashing state. With this change, the crash will prevent the
revalidation of the entry, and the next run will force an update of the
application.

R=etiennej@chromium.org

Review URL: https://codereview.chromium.org/1351693007 .